requirements.txt
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.2.2
flask-saml2==0.3.0
greenlet==2.0.2
iso8601==0.1.16
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
openpyxl==2.4.11
pyaml==21.10.1
pycparser==2.21
pymssql==2.2.7
pyOpenSSL==17.5.0
pytz==2022.2.1
PyYAML==6.0
records==0.5.3
signxml==2.10.0
six==1.16.0
SQLAlchemy==2.0.13
tablib==3.4.0
typing_extensions==4.5.0
Werkzeug==2.2.2
